Shuffle types, improve documentation
