+++ govcd/nsxt_manager.go
@@ -5,8 +5,6 @@
 package govcd
 
 import (
-	"net/url"
-
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
 )
 
@@ -18,94 +16,4 @@ type NsxtManager struct {
 	//
 	// Note:  this is being computed when retrieving the structure and will not be populated if this structure is initialized manually
 	Urn string
-}
-
-// GetAllIpDiscoveryProfiles retrieves all IP Discovery Profiles configured on an NSX-T manager.
-// NSX-T manager ID (nsxTManagerRef.id), Org VDC ID (orgVdcId) or VDC Group ID (vdcGroupId) must be
-// supplied as a filter. Results can also be filtered by a single profile ID
-// (filter=nsxTManagerRef.id==nsxTManagerUrn;id==profileId).
-func (vcdClient *VCDClient) GetAllIpDiscoveryProfiles(queryParameters url.Values) ([]*types.NsxtSegmentProfileTemplateIpDiscovery, error) {
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointNsxtSegmentIpDiscoveryProfiles
-	return genericGetAllBareFilteredEntities[types.NsxtSegmentProfileTemplateIpDiscovery](&vcdClient.Client, endpoint, queryParameters, "IP Discovery Profiles")
-}
-
-func (vcdClient *VCDClient) GetIpDiscoveryProfileByName(name string, queryParameters url.Values) (*types.NsxtSegmentProfileTemplateIpDiscovery, error) {
-	apiFilteredEntities, err := vcdClient.GetAllIpDiscoveryProfiles(queryParameters)
-	if err != nil {
-		return nil, err
-	}
-
-	return genericLocalFilterOneOrError(apiFilteredEntities, "DisplayName", name)
-}
-
-// GetAllMacDiscoveryProfiles retrieves all MAC Discovery Profiles configured on an NSX-T manager.
-// NSX-T manager ID (nsxTManagerRef.id), Org VDC ID (orgVdcId) or VDC Group ID (vdcGroupId) must be
-// supplied as a filter. Results can also be filtered by a single profile ID
-// (filter=nsxTManagerRef.id==nsxTManagerUrn;id==profileId).
-func (vcdClient *VCDClient) GetAllMacDiscoveryProfiles(queryParameters url.Values) ([]*types.NsxtSegmentProfileTemplateMacDiscovery, error) {
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointNsxtSegmentMacDiscoveryProfiles
-	return genericGetAllBareFilteredEntities[types.NsxtSegmentProfileTemplateMacDiscovery](&vcdClient.Client, endpoint, queryParameters, "MAC Discovery Profiles")
-}
-
-func (vcdClient *VCDClient) GetMacDiscoveryProfileByName(name string, queryParameters url.Values) (*types.NsxtSegmentProfileTemplateMacDiscovery, error) {
-	apiFilteredEntities, err := vcdClient.GetAllMacDiscoveryProfiles(queryParameters)
-	if err != nil {
-		return nil, err
-	}
-
-	return genericLocalFilterOneOrError(apiFilteredEntities, "DisplayName", name)
-}
-
-// GetAllSpoofGuardProfiles retrieves all Spoof Guard Profiles configured on an NSX-T manager.
-// NSX-T manager ID (nsxTManagerRef.id), Org VDC ID (orgVdcId) or VDC Group ID (vdcGroupId) must be
-// supplied as a filter. Results can also be filtered by a single profile ID
-// (filter=nsxTManagerRef.id==nsxTManagerUrn;id==profileId).
-func (vcdClient *VCDClient) GetAllSpoofGuardProfiles(queryParameters url.Values) ([]*types.NsxtSegmentProfileTemplateSegmentSpoofGuard, error) {
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointNsxtSegmentSpoofGuardProfiles
-	return genericGetAllBareFilteredEntities[types.NsxtSegmentProfileTemplateSegmentSpoofGuard](&vcdClient.Client, endpoint, queryParameters, "Spoof Guard Profiles")
-}
-
-func (vcdClient *VCDClient) GetSpoofGuardProfileByName(name string, queryParameters url.Values) (*types.NsxtSegmentProfileTemplateSegmentSpoofGuard, error) {
-	apiFilteredEntities, err := vcdClient.GetAllSpoofGuardProfiles(queryParameters)
-	if err != nil {
-		return nil, err
-	}
-
-	return genericLocalFilterOneOrError(apiFilteredEntities, "DisplayName", name)
-}
-
-// GetAllQoSProfiles retrieves all QoS Profiles configured on an NSX-T manager.
-// NSX-T manager ID (nsxTManagerRef.id), Org VDC ID (orgVdcId) or VDC Group ID (vdcGroupId) must be
-// supplied as a filter. Results can also be filtered by a single profile ID
-// (filter=nsxTManagerRef.id==nsxTManagerUrn;id==profileId).
-func (vcdClient *VCDClient) GetAllQoSProfiles(queryParameters url.Values) ([]*types.NsxtSegmentProfileTemplateSegmentQosProfile, error) {
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointNsxtSegmentQosProfiles
-	return genericGetAllBareFilteredEntities[types.NsxtSegmentProfileTemplateSegmentQosProfile](&vcdClient.Client, endpoint, queryParameters, "QoS Profiles")
-}
-
-func (vcdClient *VCDClient) GetQoSProfileByName(name string, queryParameters url.Values) (*types.NsxtSegmentProfileTemplateSegmentQosProfile, error) {
-	apiFilteredEntities, err := vcdClient.GetAllQoSProfiles(queryParameters)
-	if err != nil {
-		return nil, err
-	}
-
-	return genericLocalFilterOneOrError(apiFilteredEntities, "DisplayName", name)
-}
-
-// GetAllSegmentSecurityProfiles retrieves all Segment Security Profiles configured on an NSX-T manager.
-// NSX-T manager ID (nsxTManagerRef.id), Org VDC ID (orgVdcId) or VDC Group ID (vdcGroupId) must be
-// supplied as a filter. Results can also be filtered by a single profile ID
-// (filter=nsxTManagerRef.id==nsxTManagerUrn;id==profileId).
-func (vcdClient *VCDClient) GetAllSegmentSecurityProfiles(queryParameters url.Values) ([]*types.NsxtSegmentProfileTemplateSegmentSecurity, error) {
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointNsxtSegmentSecurityProfiles
-	return genericGetAllBareFilteredEntities[types.NsxtSegmentProfileTemplateSegmentSecurity](&vcdClient.Client, endpoint, queryParameters, "Segment Security Profiles")
-}
-
-func (vcdClient *VCDClient) GetSegmentSecurityProfileByName(name string, queryParameters url.Values) (*types.NsxtSegmentProfileTemplateSegmentSecurity, error) {
-	apiFilteredEntities, err := vcdClient.GetAllSegmentSecurityProfiles(queryParameters)
-	if err != nil {
-		return nil, err
-	}
-
-	return genericLocalFilterOneOrError(apiFilteredEntities, "DisplayName", name)
 }
